+++ network/data/src/main/kotlin/me/proton/core/network/data/ApiManagerFactory.kt
@@ -193,6 +193,8 @@ class ApiManagerFactory(
             networkManager,
             pinningStrategy,
             ::javaWallClockMs,
+            prefs,
+            cookieStore,
             extraHeaderProvider,
         )
 
@@ -223,6 +225,8 @@ class ApiManagerFactory(
                 networkManager,
                 alternativePinningStrategy,
                 ::javaWallClockMs,
+                prefs,
+                cookieStore,
                 extraHeaderProvider,
             )
         }