+++ CHANGELOG.md
@@ -30,6 +30,26 @@ with(authOrchestrator) {
 - After login, when migrating from external to internal account type, `UserCheck` is also called.
 - Fix payment refresh state after upgrading for logged in users.
 
+### Fixes
+
+- Human Verification: session cookie is now stored when DoH is enabled.
+- Human Verification: Human Verification v3 is now compatible with DoH.
+- Network: fixed a bug related to request timeouts that prevented DoH from working most of the time.
+
+### Breaking changes
+
+- `NetworkRequestOverriderImpl` now needs a `Context` parameter to be able to read the self-signed certificates.
+
+### Fixes
+
+- Human Verification: session cookie is now stored when DoH is enabled.
+- Human Verification: Human Verification v3 is now compatible with DoH.
+- Network: fixed a bug related to request timeouts that prevented DoH from working most of the time.
+
+### Breaking changes
+
+- `NetworkRequestOverriderImpl` now needs a `Context` parameter to be able to read the self-signed certificates.
+
 ## [4.0.0]
 
 ### Behavior changes